Update Turkish translations for 18.2 (by Gokturk)
"Kill" is translated as "intihar et" by the other translators. "İntihar et" means "suicide".

1. Not using the word "intihar et" because of its meaning is more accurate.
2. In game, "intihar et" causes the font getting smaller and it does not fit in the box. (https://prnt.sc/4hxQlZdoN1MS)

Using "öl" would be better instead of "intihar et" because it fits in the box better and it keeps the original meaning more.

---

The word "Name" is translated as "Takma ad", which means nickname. It causes a mistranslation by showing up above the server's name. (https://prnt.sc/1NUC5_5qR_H8)

In this scenario, using "sunucu ismi" which means "server name" would fit better.

---

Even the word "Website" means "internet sitesi" in Turkish, it does not fit in the box and can cause the font getting smaller so it is better leaving it as "web sitesi" which also used as "website" in Turkish. (https://prnt.sc/Ngkh6nkkWQF3)